- avoid needing to manually fix up the product of the templates,   fixed a bunch of ctr: directives   implemented a more precise check in __int__ (could be that such a strict check was implemented only in later   versions of cpython)
- add a basestring.expose

- make str/unicode/tuple.expose more similar to list.expose. This avoids infinite recursions when subclassing
  and overriding __getitem__ as shown by test_descr slices.
  It should not be necessary btw for Py*Derived classes to override __getitem__.

tweaking PyDictonary and commenting out a test we get to the end of test_descr now, without stack overflow explosions.

Needs to be ported to the 2.3 branch, also there we should avoid to manually tweak the template product.
We will need to change the derived templates to implement slots so just running gderived should do the job
otherwise is very unpleasant.