Update the failure and skip lists for the current state on trunk
Add a new mode, --expected.  It only runs the tests that aren't expected to
fail or be skipped such that the huge amounts of failure spew won't depress
otherwise indomitable developers.

Rip out all of the commented out single test mode stuff

When running skipped and failed tests, print out if a test passes or runs
unexpectedly in addition to the inverse.