Add an ant task to compile proxies
This works by naively importing the modules with the proxy output directory set to the destination
configured in the ant task.  Eventually it's going to need to work with our Lib directory to
generate proxies for all of our Java subclasses, but it currently fails miserably on various modules
that aren't expecting to be straight-up imported.